Initial import of ICB library and client
It's certainly not complete, however the initial scaffolding and design
should be mostly there. A Server and Client struct are exposed with
functions to handle the initial communication with a remote server.
At least enough is implemented at this point where it can login and join
the initial group, as well as receive messages.