rework how data is read from the network socket
before this change it would read a full buffer's worth of data,
look at the packet size as indicated by the very first byte in the
buffer and copy that many bytes for future parsing.

this would miss certain packets where the server would send them right
after eachother. as could be the case with a "login ok" and "status"
message. the effect was that the status message would get dropped as
it fell outside of what the parser initially thought the packet would
be.

the new implementation peeks at the socket and reads a full packet worth
of data from the socket, parses it, and returns. upon the next call to
server::read() it would parse the next message.

We now receive messages like this upon connecting to the server:
"You are now in group 1"

Also fix a typo and fmt while here.