Do not set the unread flag on a channel when adding our own messages.
Because we no longer have Tick driving the UI, the unread flag stays
on the current window until we touch it. This is nice when others
send messages, but wrong when we append our own message. So separate
out adding a message and adding a 'read' message, and use the 'read'
version for our own messages, since we have obviously read our own
messages.